+++ crates/metadata/src/lib.rs
@@ -64,6 +64,7 @@ use scale_info::{
     PortableRegistry,
     Registry,
 };
+use schemars::JsonSchema;
 use serde::{
     Deserialize,
     Serialize,
@@ -79,7 +80,7 @@ use serde::{
 /// Versions other than the `Default` are considered deprecated. If you want to
 /// deserialize legacy metadata versions you will need to use an old version of
 /// this crate.
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, JsonSchema)]
 pub enum MetadataVersion {
     #[serde(rename = "4")]
     V4,
@@ -92,7 +93,7 @@ impl Default for MetadataVersion {
 }
 
 /// An entire ink! project for metadata file generation purposes.
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, JsonSchema)]
 pub struct InkProject {
     version: MetadataVersion,
     #[serde(flatten)]